refactor: presence online-users 파일 구조 정리
