+++ src/pages/waiting-room/WaitingRoomFlow.test.tsx
@@ -26,13 +26,16 @@ vi.mock('../lobby/hooks', () => ({
   useLobbyRoomsQuery: useLobbyRoomsQueryMock,
 }))
 
-vi.mock('../../features/presence/useOnlineUsersSocket', () => ({
+vi.mock('../../features/presence/online-users/useOnlineUsersSocket', () => ({
   useOnlineUsersSocket: useOnlineUsersSocketMock,
 }))
 
-vi.mock('../../features/presence/useDirectMessageController', () => ({
-  useDirectMessageController: useDirectMessageControllerMock,
-}))
+vi.mock(
+  '../../features/presence/direct-message/useDirectMessageController',
+  () => ({
+    useDirectMessageController: useDirectMessageControllerMock,
+  })
+)
 
 // game route 도착 여부 확인용 스텁 페이지
 function GamePageStub() {
