+++ src/pages/waiting-room/WaitingRoomPage.tsx
@@ -16,10 +16,10 @@ import {
 } from '../../components/header/profileMenu'
 import { DirectMessagePanel } from '../../features/presence/components/DirectMessagePanel'
 import { UserListPanel } from '../../features/presence/components/UserListPanel'
-import { mergeOnlineUsersWithRoomPlayers } from '../../features/presence/onlineUsersModel'
-import { useOnlineUsersSocket } from '../../features/presence/useOnlineUsersSocket'
-import { useDirectMessageController } from '../../features/presence/useDirectMessageController'
-import { removeMockOnlineUser } from '../../features/presence/mockData'
+import { useDirectMessageController } from '../../features/presence/direct-message/useDirectMessageController'
+import { removeMockOnlineUser } from '../../features/presence/mock/mockData'
+import { mergeOnlineUsersWithRoomPlayers } from '../../features/presence/online-users/onlineUsersModel'
+import { useOnlineUsersSocket } from '../../features/presence/online-users/useOnlineUsersSocket'
 import { cn } from '../../lib/utils'
 import { disconnectSocketAndClearAuth } from '../../lib/socket'
 import { IS_SOCKET_MOCK_ENABLED } from '../../config/env'
