+++ src/pages/waiting-room/controller/socketSync.test.ts
@@ -33,7 +33,7 @@ vi.mock('../api', () => ({
   mapWaitingRoomSnapshotPayload: mapWaitingRoomSnapshotPayloadMock,
 }))
 
-vi.mock('../../../features/presence/onlineUsersSocket', () => ({
+vi.mock('../../../features/presence/online-users/onlineUsersSocket', () => ({
   requestOnlineUsersSnapshotSync: requestOnlineUsersSnapshotSyncMock,
 }))
